🔨 refactor row series labels
